just reorganise old template components
